Rename split_sentances to split_sentences
